+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ### 0.0.2 (SNAPSHOT)
 * Display usage on arg-less comamand
 * Upgrade vows to v0.6.1
+* Use cly to parse CLI arguments
 
 ### 0.0.1
 * Initial version